refactor: split the decomposition status rollup and context out of models
