+++ src/synthorg/engine/coordination/_middleware_relay.py
@@ -24,8 +24,8 @@ from synthorg.engine.coordination.models import (
 )
 from synthorg.engine.decomposition.models import (
     DecompositionResult,
-    SubtaskStatusRollup,
 )
+from synthorg.engine.decomposition.status_rollup import SubtaskStatusRollup
 from synthorg.engine.routing.models import RoutingResult
 
 if TYPE_CHECKING:
