+++ src/synthorg/engine/decomposition/__init__.py
@@ -5,6 +5,7 @@ classifies task structure, and manages status rollup.
 """
 
 from synthorg.engine.decomposition.classifier import TaskStructureClassifier
+from synthorg.engine.decomposition.context import DecompositionContext
 from synthorg.engine.decomposition.dag import DependencyGraph
 from synthorg.engine.decomposition.llm import (
     LlmDecompositionConfig,
@@ -12,15 +13,14 @@ from synthorg.engine.decomposition.llm import (
 )
 from synthorg.engine.decomposition.manual import ManualDecompositionStrategy
 from synthorg.engine.decomposition.models import (
-    DecompositionContext,
     DecompositionPlan,
     DecompositionResult,
     SubtaskDefinition,
-    SubtaskStatusRollup,
 )
 from synthorg.engine.decomposition.protocol import DecompositionStrategy
 from synthorg.engine.decomposition.rollup import StatusRollup
 from synthorg.engine.decomposition.service import DecompositionService
+from synthorg.engine.decomposition.status_rollup import SubtaskStatusRollup
 
 __all__ = [
     "DecompositionContext",
