+++ src/synthorg/engine/decomposition/protocol.py
@@ -4,10 +4,8 @@ from typing import Protocol, runtime_checkable
 
 from synthorg.core.task import Task
 from synthorg.core.types import NotBlankStr
-from synthorg.engine.decomposition.models import (
-    DecompositionContext,
-    DecompositionPlan,
-)
+from synthorg.engine.decomposition.context import DecompositionContext
+from synthorg.engine.decomposition.models import DecompositionPlan
 
 
 @runtime_checkable
