+++ src/synthorg/engine/middleware/coordination_protocol.py
@@ -21,8 +21,8 @@ from synthorg.engine.coordination.models import (
 )
 from synthorg.engine.decomposition.models import (
     DecompositionResult,
-    SubtaskStatusRollup,
 )
+from synthorg.engine.decomposition.status_rollup import SubtaskStatusRollup
 from synthorg.engine.middleware.models import TaskLedger
 from synthorg.engine.routing.models import RoutingResult
 from synthorg.observability import get_logger, safe_error_description
