+++ tests/unit/engine/middleware/test_coordination_constraints.py
@@ -17,8 +17,8 @@ from synthorg.engine.coordination.models import (
 )
 from synthorg.engine.decomposition.models import (
     DecompositionResult,
-    SubtaskStatusRollup,
 )
+from synthorg.engine.decomposition.status_rollup import SubtaskStatusRollup
 from synthorg.engine.middleware.coordination_constraints import (
     PlanReviewGateMiddleware,
     TaskLedgerMiddleware,
