+++ tests/unit/engine/test_decomposition_llm_prompt.py
@@ -19,6 +19,7 @@ from synthorg.core.task_enums import (
     TaskType,
 )
 from synthorg.core.types import NotBlankStr
+from synthorg.engine.decomposition.context import DecompositionContext
 from synthorg.engine.decomposition.llm_parse import (
     parse_content_response,
     parse_tool_call_response,
@@ -29,10 +30,7 @@ from synthorg.engine.decomposition.llm_prompt import (
     build_system_message,
     build_task_message,
 )
-from synthorg.engine.decomposition.models import (
-    DecompositionContext,
-    DecompositionPlan,
-)
+from synthorg.engine.decomposition.models import DecompositionPlan
 from synthorg.engine.errors import DecompositionError
 from synthorg.providers.enums import MessageRole
 from synthorg.providers.models import (
