+++ tests/unit/engine/test_parent_rollup.py
@@ -20,13 +20,10 @@ from synthorg.engine.coordination.parent_rollup import (
     run_update_parent_phase,
 )
 from synthorg.engine.decomposition.classifier import TaskStructureClassifier
-from synthorg.engine.decomposition.models import (
-    DecompositionContext,
-    DecompositionPlan,
-    DecompositionResult,
-    SubtaskStatusRollup,
-)
+from synthorg.engine.decomposition.context import DecompositionContext
+from synthorg.engine.decomposition.models import DecompositionPlan, DecompositionResult
 from synthorg.engine.decomposition.service import DecompositionService
+from synthorg.engine.decomposition.status_rollup import SubtaskStatusRollup
 from synthorg.engine.task_engine import TaskEngine
 from synthorg.engine.task_engine_models import TaskMutationResult
 from tests._shared import FakeClock, mock_of
